// src/logic/actions/action.types.ts
export enum ActionTypes {
  INITIALISING = 'INITIALISING',
  BEGIN_PROCESSING_INPUTS = 'BEGIN_PROCESSING_INPUTS',
  CSV_PROCESSING_START = 'CSV_PROCESSING_START',
  CSV_PROCESSING_SUCCESS = 'CSV_PROCESSING_SUCCESS',
  CSV_PROCESSING_ERROR = 'CSV_PROCESSING_ERROR',
  POCKETSMITH_FETCH_START = 'POCKETSMITH_FETCH_START',
  POCKETSMITH_FETCH_SUCCESS = 'POCKETSMITH_FETCH_SUCCESS',
  POCKETSMITH_FETCH_ERROR = 'POCKETSMITH_FETCH_ERROR',
  TRANSACTION_MATCHING_START = 'TRANSACTION_MATCHING_START',
  TRANSACTION_MATCHING_SUCCESS = 'TRANSACTION_MATCHING_SUCCESS',
  TRANSACTION_MATCHING_ERROR = 'TRANSACTION_MATCHING_ERROR',
  CONFIRMING_MATCHED_TRANSACTIONS_START = 'CONFIRMING_MATCHED_TRANSACTIONS_START',
  SET_CURRENT_MATCH_INDEX = 'SET_CURRENT_MATCH_INDEX',
  CONFIRM_MATCH = 'CONFIRM_MATCH',
  REJECT_MATCH = 'REJECT_MATCH',
  UPDATE_POCKETSMITH_TRANSACTION = 'UPDATE_POCKETSMITH_TRANSACTION',
  CONFIRMING_MATCHED_TRANSACTIONS_COMPLETE = 'CONFIRMING_MATCHED_TRANSACTIONS_COMPLETE',
  INTERACTIVE_MATCHING_START = 'INTERACTIVE_MATCHING_START',
  SET_CURRENT_UNMATCHED_TRANSACTION = 'SET_CURRENT_UNMATCHED_TRANSACTION',
  USER_RESPONSE_FOUND_MATCH = 'USER_RESPONSE_FOUND_MATCH',
  USER_RESPONSE_NO_MATCH = 'USER_RESPONSE_NO_MATCH',
  SET_MANUAL_MATCH_DETAILS = 'SET_MANUAL_MATCH_DETAILS',
  SET_MANUAL_MULTI_MATCH_DETAILS = 'SET_MANUAL_MULTI_MATCH_DETAILS',
  PROCESS_MANUAL_MATCH = 'PROCESS_MANUAL_MATCH',
  PAYPAL_ID_CONFLICT_DETECTED = 'PAYPAL_ID_CONFLICT_DETECTED',
  RESOLVE_PAYPAL_ID_CONFLICT = 'RESOLVE_PAYPAL_ID_CONFLICT',
  INTERACTIVE_MATCHING_COMPLETE = 'INTERACTIVE_MATCHING_COMPLETE',
  PROCESSING_COMPLETE = 'PROCESSING_COMPLETE',
  FETCH_TRANSACTION_ACCOUNTS = 'FETCH_TRANSACTION_ACCOUNTS',
  FETCH_TRANSACTION_ACCOUNTS_SUCCESS = 'FETCH_TRANSACTION_ACCOUNTS_SUCCESS',
  FETCH_TRANSACTION_ACCOUNTS_ERROR = 'FETCH_TRANSACTION_ACCOUNTS_ERROR',
  SELECT_TRANSACTION_ACCOUNT = 'SELECT_TRANSACTION_ACCOUNT',
  SKIP_ACCOUNT_SELECTION = 'SKIP_ACCOUNT_SELECTION',
  MATCHING_AMAZON_TRANSACTIONS = 'MATCHING_AMAZON_TRANSACTIONS',
  AMAZON_TRANSACTION_MATCHING_SUCCESS = 'AMAZON_TRANSACTION_MATCHING_SUCCESS',
  AMAZON_TRANSACTION_MATCHING_ERROR = 'AMAZON_TRANSACTION_MATCHING_ERROR',
}

export type Action = {
  type: ActionTypes;
  payload: unknown;
};